requirements.txt
numpy==1.21.6
pandas==1.3.4
matplotlib==3.5.3
seaborn==0.12.1
keras==2.7.0
scikit-learn==1.0.2
tensorflow==2.7.0
streamlit==1.17.0
streamlit-drawable-canvas==0.9.2
scikit-image==0.19.3
gensim==4.1.1
Pillow


